change button styles
