+++ src/components/Link.js
@@ -4,10 +4,10 @@ import Button from '@material-ui/core/Button'
 import { StyleSheet, css } from 'aphrodite';
 
 const Link = ({ active, children, onClick}) => (
-  <Button className={css(styles.zippyHeader)}
+  <Button className={css(styles.hover)}
     onClick={onClick}
     disabled={active}
-    style={{
+    style={active ? {color: 'black',backgroundColor: '#FF80AB',width:"100px",height:"10px",}:{
       color: 'white',
       backgroundColor: 'lightGreen',
       marginLeft: '1px',
@@ -27,38 +27,11 @@ Link.propTypes = {
 
 export default Link
 
-const translateKeyframes = {
-    '0%': {
-        transform: 'translateX(0)',
-    },
-
-    '50%': {
-        transform: 'translateX(100px)',
-    },
-
-    '100%': {
-        transform: 'translateX(0)',
-    },
-};
-
-const opacityKeyframes = {
-    'from': {
-        opacity: 0,
-    },
-
-    'to': {
-        opacity: 1,
-    }
-};
-
 
 const styles = StyleSheet.create({
-  zippyHeader: {
+  hover: {
     ':hover': {
         backgroundColor: 'red'
     },
-    animationName: [translateKeyframes, opacityKeyframes],
-    animationDuration: '3s, 1200ms',
-    animationIterationCount: 'infinite',
   },
 });
