+++ src/index.js
@@ -7,23 +7,60 @@ import App from './components/App'
 
 import { createStore, applyMiddleware} from "redux"
 import reduxThunk from "redux-thunk";
+
+
+import Button from '@material-ui/core/Button'
+import { StyleSheet, css } from 'aphrodite';
+
 const store = createStore(rootReducer, {}, applyMiddleware(reduxThunk));
 
+const translateKeyframes = {
+    '0%': {
+        transform: 'translateX(0)',
+    },
+
+    '50%': {
+        transform: 'translateX(100px)',
+    },
+
+    '100%': {
+        transform: 'translateX(0)',
+    },
+};
+
+const opacityKeyframes = {
+    'from': {
+        opacity: 0,
+    },
+
+    'to': {
+        opacity: 1,
+    }
+};
+
+
+const styles = StyleSheet.create({
+  zippyHeader: {
+    'margin-top': '50px',
+    ':hover': {
+        backgroundColor: 'red'
+    },
+    animationName: [translateKeyframes, opacityKeyframes],
+    animationDuration: '3s, 1200ms',
+    animationIterationCount: 'infinite',
+  },
+});
 
 ReactDOM.render(
   <div>
     <Provider store={store}>
       <App />
     </Provider>
-    <button onClick={() => console.log(store.getState())}
-            className="test"
+    <Button className={css(styles.zippyHeader)}
+    onClick={() => console.log(styles.zippyHeader)}
     >
-    Current Store tree
-    </button>
+    Redux Store tree
+    </Button>
   </div>,
    document.getElementById('root')
 );
-
-// If you want your app to work offline and load faster, you can change
-// unregister() to register() below. Note this comes with some pitfalls.
-// Learn more about service workers: http://bit.ly/CRA-PWA
